fix: address second round of reviews
