+++ riocli/config/config.py
@@ -16,8 +16,10 @@ from __future__ import annotations
 import errno
 import json
 import os
+import subprocess
 import uuid
 from functools import lru_cache
+from pathlib import Path
 
 from click import get_app_dir
 from rapyuta_io import Client
@@ -211,3 +213,83 @@ class Configuration:
             self.save()
 
         return self.data.get("machine_id")
+
+    # ------------------------------------------------------------------
+    # SSH key pair management
+    # ------------------------------------------------------------------
+
+    #: Base name for the dedicated SSH key pair.
+    _SSH_KEY_NAME = "rio_ed25519"
+
+    @property
+    def ssh_private_key(self: Configuration) -> Path:
+        """Path to the dedicated SSH private key (``rio_ed25519``).
+
+        Located in the CLI's configuration directory alongside
+        ``config.json``.
+        """
+        return Path(get_app_dir(self.APP_NAME)) / self._SSH_KEY_NAME
+
+    @property
+    def ssh_public_key(self: Configuration) -> Path:
+        """Path to the dedicated SSH public key (``rio_ed25519.pub``)."""
+        return Path(get_app_dir(self.APP_NAME)) / f"{self._SSH_KEY_NAME}.pub"
+
+    @property
+    def ssh_certificate(self: Configuration) -> Path:
+        """Path to the SSH certificate (``rio_ed25519-cert.pub``)."""
+        return Path(get_app_dir(self.APP_NAME)) / f"{self._SSH_KEY_NAME}-cert.pub"
+
+    def ensure_ssh_keys(self: Configuration) -> bool:
+        """Ensure the dedicated ed25519 key pair exists, creating if needed.
+
+        If both the private and public key files exist, they are reused.
+        Otherwise, any orphaned files (including a stale certificate) are
+        cleaned up and a fresh key pair is generated via ``ssh-keygen``.
+
+        Returns:
+            ``True`` when a new key pair was generated, ``False`` when
+            the existing pair was reused.
+        """
+        private_path = self.ssh_private_key
+        public_path = self.ssh_public_key
+        cert_path = self.ssh_certificate
+
+        # Ensure the config directory exists with correct permissions.
+        private_path.parent.mkdir(mode=0o700, parents=True, exist_ok=True)
+
+        if private_path.is_file() and public_path.is_file():
+            return False
+
+        # Orphaned files from a previous key pair must be removed to
+        # avoid cert/key mismatch and because ssh-keygen refuses to
+        # overwrite an existing private key.
+        private_path.unlink(missing_ok=True)
+        public_path.unlink(missing_ok=True)
+        cert_path.unlink(missing_ok=True)
+
+        # Generate a fresh ed25519 key pair.
+        result = subprocess.run(
+            [
+                "ssh-keygen",
+                "-t",
+                "ed25519",
+                "-f",
+                str(private_path),
+                "-N",
+                "",
+                "-C",
+                "rio-ssh",
+                "-q",
+            ],
+            capture_output=True,
+            text=True,
+        )
+        if result.returncode != 0:
+            raise RuntimeError(f"ssh-keygen failed: {result.stderr.strip()}")
+
+        # Ensure correct permissions (ssh-keygen normally sets these).
+        private_path.chmod(0o600)
+        public_path.chmod(0o644)
+
+        return True
